adds check to methods start and abort for navigator.onLine
the browser behavior appears to remove all outstanding async requests
once it sees the first request failed due to connection error. this
means that other requests are not properly aborted in a way that
activates ProgressIndicator code to decrement activeReqs. So adding the
check for navigator.onLine allows ProgressIndicator to know when all
progress should be stoppend due to connection error.