updates .travis.yml to upload with cache_control
note: the /dist folder is uploaded a second time with different
cache_control settings. this is a work-a-round for an issue with travis
dpl where it is only able to upload directories and not individual
files. This means that cache_control can only be set on a whole folder
basis.