add max-width 100% for all children of Post body
this change was necessary because images wider than the available width
of the body area would overflow. in order to keep these images within
the margins of the body area, a css rule was added which sets the
max-width to 100% for all children elements which causes images that are
too wide to autosize to fit the width.